Creates an object individual

Creates an object individual

individual(
  X,
  y,
  clf,
  coeffs = NULL,
  ind = NULL,
  eval.all = FALSE,
  signs = NULL,
  obj = NULL,
  res_clf = NULL
)

Arguments

X:

the data matrix with variables in the rows and observations in the columns

y:

the class vector

clf:

the object containing the classifier information

ind:

the indexes of the variables forming the individual could be null if we give the function a dense vector (via the coeff parameter) or if we also want to generate the individual

coeffs:

the coefficients of the model, it could be a dense vector (in this case, ind need to be null), or it could be only the non zero values, or if it's null a new individual will be genrated

obj:

an object to be incorporated in the model (default:NULL). We use this usually for SOTA.

res_clf:

if provided information on mda etc can be found and transmitted to the model object

Value

an individual (model) object