mergeMeltImportanceCV

mergeMeltImportanceCV returns a list of data frames that contain the feature importance of the different learners without any focus on sparsity.

mergeMeltImportanceCV(
  list.results,
  filter.cv.prev = 0.5,
  min.kfold.nb = FALSE,
  type = "mda",
  learner.grep.pattern = "*",
  nb.top.features = 25,
  feature.selection = NULL,
  fixed.order = FALSE,
  scaled.importance = TRUE,
  make.plot = TRUE,
  main = FALSE,
  cv.prevalence = TRUE
)

Arguments

list.results.digest:

a list of digest objects one for each learner used. For example, list(res.terda.digest, res.terga.digest, res.terbeam.digest)

filter.cv.prev:

filter variable for each learner based on the appearence prevalence in the cross validation.

min.kfold.nb:

wether we should restrict all experiments in the smallest number of k-folds of a comparative analyses (default = FALSE)

type:

the type of importance "mda (mean decreased accuracy)" or "pda (prevalence decreased accuracy)" (default = mda)

learner.grep.pattern:

select a subset of learners using a grep pattern (default:"*")

nb.top.features:

the number of top features to focus on the plot

feature.selection:

the names of the features to be selected (default:NULL)

fixed.order:

if the order of features in the plot should follow the feature selection one (default = FALSE)

scaled.importance:

the scaled importance is the importance multipied by the prevalence in the folds. If (default = TRUE) this will be used, the mean mda will be scaled by the prevalence of the feature in the folds and ordered subsequently

make.plot:

make a plot for all the learners

main:

should add the title to the graph for correct alignment (default:FALSE)

cv.prevalence:

wether or not to plot the distribution of the prevalence of the feature in the top-models for each k-fold in the graph (default:FALSE)

Value

a list of several data.frames and a ggplot object

Details

Merge a list of cross validation scores form digest results