sota.svm

sota.svm is a wrapper that executes svm using the same framework as for the predomics package.

sota.svm(
  sparsity = c(1:30),
  objective = "auc",
  max.nb.features = 1000,
  intercept = 0,
  language = "svm",
  evalToFit = "auc_",
  k_penalty = 0,
  scaled = TRUE,
  type = NULL,
  kernel = "rbfdot",
  kpar = "automatic",
  C = c(1e-04, 0.001, 0.01, 0.1, 1, 10, 100, 1000, 10000),
  nu = 0.2,
  epsilon.hp = 0.1,
  prob.model = FALSE,
  class.weights = NULL,
  fit = TRUE,
  cache = 40,
  tol = 0.001,
  shrinking = TRUE,
  na.action = na.omit,
  popSaveFile = "NULL",
  seed = "NULL",
  nCores = 4,
  verbose = TRUE,
  plot = FALSE,
  warnings = FALSE,
  debug = FALSE,
  print_ind_method = "short",
  experiment.id = NULL,
  experiment.description = NULL,
  experiment.save = "nothing"
)

Arguments

language

is the language that is used by the different algorithms bin, bininter, ter, terinter, ratio, (default:"sota")

sparsity:

number of features in a given model. This is a vector with multiple lengths.

objective:

prediction mode (default: auc)

max.nb.features:

create the glmnet object using only the top most significant features (default:1000)

intercept:

(Interceot for the a given model) (default:NULL)

evalToFit:

Which model property will be used to select the best model among different k_sparsities (default: auc_)

k_penalty:

Penalization of the fit by the k_sparsity (default: 0)

scaled:

??

type:

??

kernel:

??

kpar:

??

C:

(??)

nu:

??

epsilon.hp:

(??) (for the SVM)

prob.model:

??

class.weights:

??

fit:

??

cache:

(??)

tol:

??

shrinking:

??

na.action:

??

popSaveFile:

(??)

seed:

the seed to be used for reproductibility. If seed=NULL than it is not taken into account (default:NULL).

nCores:

the number of CPUs to run the program in parallel

plot:

Plot graphics indicating the evolution of the simulation (default:FALSE)

verbose:

print out information on the progress of the algorithm (default:TRUE)

warnings:

Print out warnings when runnig (default:FALSE).

debug:

print out information on the progress of the algorithm (default:FALSE)

print_ind_method:

One of c("short","graphical") indicates how to print a model and subsequently a population during the run (default:"short").

experiment.id:

The id of the experiment that is to be used in the plots and comparitive analyses (default is the learner's name, when not specified)

experiment.description:

A longer description of the experiment. This is important when many experiments are run and can also be printed in by the printExperiment function.

experiment.save:

Data from an experiment can be saved with different levels of completness, with options to be selected from c("nothing", "minimal", "full"), default is "minimal"

Value

an object containing a list of parameters for this classifier

Details

sota.svm: launching svm classifier